+++ src/lib/types.ts
@@ -1,7 +1,17 @@
 export type Memo = {
   name: string;
   uid: string;
+  pinned: boolean;
+  rowStatus: string;
   content: string;
+  tags: string[];
+  creator: string;
+  createTime: string;
+  updateTime: string;
+};
+
+export type MemoWithSummary = Memo & {
+  summary: string;
 };
 
 export interface MemoListProps {